Add SetUnion extension from the paper
So far, optimizations won't work, since we'll need to declared instances of
SymanticsS for each pass (fortunately we won't need to implement anything,
thanks to the default method implementation, but it is still annoying we have to
declare all the instances).

Alternatively we could add a generic
(Symantics repr, RR t repr) => SymanticsS (t repr) instance to automatically
capture all passes that don't have to deal with the set union.

But that would mean overlapping instance in case we would like to implement
optimization pass that cares about set union...

I don't know which option is better...